+++ src/Exception/ConfigExceptionInterface.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * RCC Accounting Webservices API
+ *
+ * @license Copyright (c) Florida State University; all rights reserved *
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ *
+ * ----------------------------------------------------------------------------
+ */
+
+declare(strict_types=1);
+
+namespace Configula\Exception;
+
+/**
+ * All exception classes in this library implement ConfigExceptionInterface
+ */
+interface ConfigExceptionInterface
+{
+    // pass...
+}